+++ wifipumpkin3/core/servers/mitm/responder3.py
@@ -55,6 +55,7 @@ class Responder3(MitmMode):
     ModSettings = True
     ModType = "server"  # proxy or server
     config = None
+    TypeButton = 0 # 0 for Switch, 1 for Radio
 
     def __init__(self, parent, FSettingsUI=None, main_method=None, **kwargs):
         super(Responder3, self).__init__(parent)