+++ wifipumpkin3/core/servers/proxy/no_proxy.py
@@ -26,6 +26,7 @@ class NoProxy(ProxyMode):
     Author = "Pumpkin-Dev"
     Hidden = True
     TypePlugin = 1
+    TypeButton = 1 # 0 for Switch, 1 for Radio
 
     def __init__(self, parent, **kwargs):
         super(NoProxy, self).__init__(parent)