+++ wifipumpkin3/core/servers/proxy/proxymode.py
@@ -47,6 +47,7 @@ class ProxyMode(Widget, ComponentBlueprint):
     _cmd_array = []
     Hidden = True
     plugins = []
+    TypeButton = 0 # 0 for Switch, 1 for Radio
     sendError = QtCore.pyqtSignal(str)
     sendSingal_disable = QtCore.pyqtSignal(object)
     addDock = QtCore.pyqtSignal(object)