+++ wifipumpkin3/core/servers/rest/blueprints/restapi/accesspoint.py
@@ -6,6 +6,7 @@ from netaddr import EUI
 from flask_restful import Resource
 from flask import jsonify, request
 from wifipumpkin3.core.servers.rest.ext.auth import token_required
+from wifipumpkin3.core.servers.rest.ext.exceptions import exception
 
 # This file is part of the wifipumpkin3 Open Source Project.
 # wifipumpkin3 is licensed under the Apache 2.0.