+++ wifipumpkin3/core/servers/rest/blueprints/restapi/commands.py
@@ -1,10 +1,10 @@
-from flask import jsonify
+from flask import jsonify, request
 from wifipumpkin3 import PumpkinShell
 from flask_restful import Resource
 from io import StringIO
 import sys
 from wifipumpkin3.core.servers.rest.ext.auth import token_required
-
+from wifipumpkin3.core.servers.rest.ext.exceptions import exception
 # This file is part of the wifipumpkin3 Open Source Project.
 # wifipumpkin3 is licensed under the Apache 2.0.
 
@@ -45,4 +45,23 @@ class CommandsResource(Resource):
         output = []
         with Capturing(output) as output:
             self.root.onecmd(command)
+        return jsonify(output)
+
+
+class CommandsPostResource(Resource):
+    def __init__(self):
+        self.root = PumpkinShell.getInstance()
+        super(CommandsPostResource, self).__init__()
+
+    @token_required
+    def post(self, command=None):
+        data = request.get_json(force=True)
+        if not "commands" in data:
+            return exception(
+                "Cannot found that key commands  on data",
+                code=400,
+            )
+        output = []
+        with Capturing(output) as output:
+            self.root.onecmd(data.get("commands"))
         return jsonify(output)