+++ wifipumpkin3/core/servers/rest/ext/auth.py
@@ -1,6 +1,6 @@
 from wifipumpkin3.core.utility.collection import SettingsINI
 import uuid
-from flask import request, jsonify
+from flask import request, jsonify, make_response
 import jwt
 from functools import wraps
 
@@ -31,7 +31,10 @@ def token_required(f):
             token = request.headers["x-access-token"]
 
         if not token:
-            return jsonify({"message": "Token is missing!"})
+            return make_response(
+                jsonify({"message": "Token is invalid!"}),
+                401
+            )
         conf = SettingsINI.getInstance()
 
         try:
@@ -39,9 +42,15 @@ def token_required(f):
             data = jwt.decode(token, app_secret_key)
             app_public_id = conf.get("rest_api_settings", "public_id")
             if app_public_id != data["public_id"]:
-                return jsonify({"message": "Token is invalid!"})
+                return make_response(
+                jsonify({"message": "Token is invalid!"}),
+                401
+            )
         except:
-            return jsonify({"message": "Token is invalid!"})
+            return make_response(
+                jsonify({"message": "Token is invalid!"}),
+                401
+            )
 
         return f(*args, **kwargs)